requirements.txt
trimesh==4.2.3
accelerate==0.28.0
mesh2sdf==1.1.0
einops==0.7.0
einx==0.1.3
optimum==1.18.0
omegaconf==2.3.0
opencv-python==4.9.0.80
transformers==4.39.3
numpy==1.26.4
huggingface_hub
matplotlib
gradio
spaces
Pillow